+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * no inflection **required**
 * bug with lexicon-less extraction
 
+### Added
+* `dictionary` command
+
 ## [0.0.2] - 2023-07-02
 
 ### Added